requirement.txt
dts==0.1.3
future==0.18.3
gpiozero==1.6.2
iso8601==2.0.0
keyboard==0.13.5
minimalmodbus==2.0.1
modbus-tk==1.1.3
numpy==1.21.6
opencv-python==4.8.0.74
pyserial==3.5
python-dateutil==2.8.2
python-periphery==2.4.1
PyYAML==6.0.1
RPi.GPIO==0.7.1
serial==0.0.97

